+++ CHANGELOG.md
@@ -15,6 +15,8 @@
   custom matchers. ([#5162](https://github.com/facebook/jest/pull/5162))
 * `[pretty-format]` Pretty format for DOMStringMap and NamedNodeMap
   ([#5233](https://github.com/facebook/jest/pull/5233))
+* `[jest-cli]` Use a better console-clearing string on Windows
+  ([#5251](https://github.com/facebook/jest/pull/5251))
 
 ### Features